# app/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "BSC-VRE repository",
        "description": "This API manages the files stored in the local data repository of a MuG Virtual Research Enviroment (VRE). It allows to display and download these files connecting to a local storage.<br/>In this implemetation, the local repository corresponds to the BSC-VRE repository, accessible via web in http://multiscalegenomics.bsc.es",
        "contact": {
            "email": "multiscalegenomics@gmail.com"
        },
        "license": {
            "name": "Apache 2.0 (?)",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "multiscalegenomics.bsc.es",
    "basePath": "/VREapi/public/v1",
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "paths": {
        "/files": {
            "get": {
                "tags": [
                    "files"
                ],
                "summary": "List user files",
                "description": "List MuG files accessible by the user. Result can be filtered and sorted",
                "operationId": "showUserMugfiles",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit number of MuG files returned",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Sort files by attribute",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Mugfile"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "401": {
                        "description": "Not authorized"
                    }
                },
                "deprecated": true,
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/files/{id}": {
            "get": {
                "tags": [
                    "files"
                ],
                "summary": "Show Mugfile",
                "description": "Show MuG file object",
                "operationId": "showMugfile",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Mugfile identifier",
                        "required": true,
                        "type": "string",
                        "default": "MuGUSER5968cc290156d_5968cc2954a382.7980548"
                    },
                    {
                        "name": "access_token",
                        "in": "query",
                        "description": "Access Token. If not found in header, accepted from URL query",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/Mugfile"
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "401": {
                        "description": "Not authorized"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/content/{file_path}": {
            "get": {
                "tags": [
                    "content"
                ],
                "summary": "Get file content",
                "description": "Get file content given a the path in the repository ('file_path' in Mugfile object)",
                "operationId": "getFileContent",
                "produces": [
                    "application/json",
                    "text/plain",
                    "application/octet-stream",
                    "application/x-gzip",
                    "application/x-tar",
                    "application/zip",
                    "text/html",
                    "image/png",
                    "image/tiff"
                ],
                "parameters": [
                    {
                        "name": "file_path",
                        "in": "path",
                        "description": "Mugfile file_path",
                        "required": true,
                        "type": "string",
                        "default": "MuGUSER5968cc290156d\\/uploads\\/README.md"
                    },
                    {
                        "name": "access_token",
                        "in": "query",
                        "description": "Access Token. If not found in header, accepted from URL query",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns content file",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    },
                    "401": {
                        "description": "Not authorized"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "definitions": {
        "Error": {
            "required": [
                "code",
                "message"
            ],
            "properties": {
                "code": {
                    "description": "HTTP code",
                    "type": "integer"
                },
                "message": {
                    "description": "Error description",
                    "type": "string"
                }
            },
            "type": "object",
            "xml": {
                "name": "Error"
            }
        },
        "Mugfile": {
            "required": [
                "file_id",
                "file_path",
                "file_type",
                "data_type",
                "taxon_id",
                "compressed",
                "source_id"
            ],
            "properties": {
                "file_id": {
                    "description": "MuG file identifier",
                    "type": "string"
                },
                "file_path": {
                    "description": "File path relative to user root directory",
                    "type": "string"
                },
                "file_type": {
                    "description": "File format",
                    "type": "string"
                },
                "data_type": {
                    "type": "string"
                },
                "taxon_id": {
                    "description": "Taxon identifier refered to the data contained in the file",
                    "type": "integer"
                },
                "compressed": {
                    "description": "Compression state of the file",
                    "type": "string"
                },
                "source_id": {
                    "description": "MuG file identifier of the source file from which the current derives",
                    "type": "string"
                },
                "meta_data": {
                    "description": "File meta data optional according the file_type and the data_type (i.e. refGenome)",
                    "type": "object"
                },
                "creation_time": {
                    "description": "Stamp time of the file creation date",
                    "type": "string"
                }
            },
            "type": "object",
            "xml": {
                "name": "Mugfile"
            }
        }
    },
    "securityDefinitions": {
        "bearer": {
            "type": "apiKey",
            "description": "This API expects users to authentificate via bearer tokens. Obtain them at http://multiscalegenomics.bsc.es,</p><p>1 - login using any of the configured OAuth2 providers and your token will be displayed at 'My Profile' &rarr; 'API Keys' section.</p><p>2 - Copy&paste it here in the following format: 'bearer access_token_string'",
            "name": "Authorization",
            "in": "header"
        },
        "mug_auth": {
            "type": "oauth2",
            "description": "This API expects users to authentificate via bearer tokens obtained via OAuth2. If you have an authorized OAuth2 client, you can generate them here. If not, open the OAuth2 dialog at http://multiscalegenomics.bsc.es to obtain them, as detailed in the above section<br/>",
            "flow": "accessCode",
            "authorizationUrl": "https://inb.bsc.es/auth/realms/mug/protocol/openid-connect/auth",
            "tokenUrl": "https://inb.bsc.es/auth/realms/mug/protocol/openid-connect/token",
            "scopes": {
                "read": "read your files"
            }
        }
    },
    "externalDocs": {
        "description": "Find out more about Swagger",
        "url": "http://swagger.io"
    }
}